+++ lib/blue_print_pos.dart
@@ -3,36 +3,55 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui';
 
-import 'package:blue_print_pos/models/batch_print_options.dart';
 import 'package:blue_print_pos/models/models.dart';
 import 'package:blue_print_pos/receipt/receipt_section_text.dart';
-import 'package:blue_print_pos/scanner/blue_scanner.dart';
-import 'package:blue_thermal_printer/blue_thermal_printer.dart' as blue_thermal;
 import 'package:esc_pos_utils_plus/esc_pos_utils.dart';
+import 'package:fluetooth/fluetooth.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_blue_plus/flutter_blue_plus.dart' as flutter_blue;
-import 'package:flutter_blue_plus/gen/flutterblueplus.pb.dart' as proto;
 import 'package:image/image.dart' as img;
 import 'package:qr_flutter/qr_flutter.dart';
 
-class BluePrintPos {
-  BluePrintPos._() {
-    _bluetoothAndroid = blue_thermal.BlueThermalPrinter.instance;
-    _bluetoothIOS = flutter_blue.FlutterBluePlus.instance;
-  }
+export 'package:esc_pos_utils_plus/esc_pos_utils.dart' show PaperSize;
+export 'package:fluetooth/fluetooth.dart' show FluetoothDevice;
+export 'models/models.dart';
+export 'receipt/receipt.dart';
 
-  static BluePrintPos get instance => BluePrintPos._();
+class BluePrintPos {
+  static final BluePrintPos _instance = BluePrintPos();
+
+  static BluePrintPos get instance => _instance;
 
   static const MethodChannel _channel = MethodChannel('blue_print_pos');
 
-  /// This field is library to handle in Android Platform
-  blue_thermal.BlueThermalPrinter? _bluetoothAndroid;
+  static final PrinterFeatures _printerFeatures = PrinterFeatures();
 
-  /// This field is library to handle in iOS Platform
-  flutter_blue.FlutterBluePlus? _bluetoothIOS;
+  /// Register printer device name with its features.
+  /// Example:
+  /// ```dart
+  /// BluePrintPos.addPrinterFeatures(<String, Set<PrinterFeature>>{
+  ///   // The name of the device is from [FluetoothDevice.name]
+  ///   const PrinterFeatureRule.allowFor('PRJ-80AT-BT'): {
+  //      PrinterFeature.paperFullCut,
+  //    },
+  ///   // Allow all printers for this feature
+  ///   PrinterFeatureRule.allowAll: {PrinterFeature.paperFullCut},
+  /// });
+  /// ```
+  ///
+  /// If [BluePrintPos.selectedDevice] name does not match, for example, for
+  /// [PrinterFeature.paperFullCut], then when printing using `useCut`
+  /// it will not produce any ESC command for paper full cut.
+  static void addPrinterFeatures(PrinterFeatureMap features) {
+    _printerFeatures.featureMap.addAll(features);
+  }
 
-  /// Bluetooth Device model for iOS
-  flutter_blue.BluetoothDevice? _bluetoothDeviceIOS;
+  /// Check if the printer has the feature.
+  ///
+  /// This will return `true` if [feature] is allowed for the printer or
+  /// if [feature] is allowed for all printers.
+  static bool printerHasFeatureOf(String printerName, PrinterFeature feature) {
+    return _printerFeatures.hasFeatureOf(printerName, feature);
+  }
 
   /// State to get bluetooth is connected
   bool _isConnected = false;
@@ -40,12 +59,14 @@ class BluePrintPos {
   /// Getter value [_isConnected]
   bool get isConnected => _isConnected;
 
+  FluetoothDevice? _selectedDevice;
+
   /// Selected device after connecting
-  BlueDevice? selectedDevice;
+  FluetoothDevice? get selectedDevice => _selectedDevice;
 
   /// return bluetooth device list, handler Android and iOS in [BlueScanner]
-  Future<List<BlueDevice>> scan() async {
-    return await BlueScanner.scan();
+  Future<List<FluetoothDevice>> scan() {
+    return Fluetooth().getAvailableDevices();
   }
 
   /// When connecting, reassign value [selectedDevice] from parameter [device]
@@ -53,61 +74,28 @@ class BluePrintPos {
   /// will return [ConnectionStatus.timeout]
   /// When connection success, will return [ConnectionStatus.connected]
   Future<ConnectionStatus> connect(
-    BlueDevice device, {
+    FluetoothDevice device, {
     Duration timeout = const Duration(seconds: 5),
   }) async {
-    selectedDevice = device;
     try {
-      if (Platform.isAndroid) {
-        final blue_thermal.BluetoothDevice bluetoothDeviceAndroid =
-            blue_thermal.BluetoothDevice(
-                selectedDevice?.name ?? '', selectedDevice?.address ?? '');
-        await _bluetoothAndroid?.connect(bluetoothDeviceAndroid);
-      } else if (Platform.isIOS) {
-        _bluetoothDeviceIOS = flutter_blue.BluetoothDevice.fromProto(
-          proto.BluetoothDevice(
-            name: selectedDevice?.name ?? '',
-            remoteId: selectedDevice?.address ?? '',
-            type: proto.BluetoothDevice_Type.valueOf(selectedDevice?.type ?? 0),
-          ),
-        );
-        final List<flutter_blue.BluetoothDevice> connectedDevices =
-            await _bluetoothIOS?.connectedDevices ??
-                <flutter_blue.BluetoothDevice>[];
-        final int deviceConnectedIndex = connectedDevices
-            .indexWhere((flutter_blue.BluetoothDevice bluetoothDevice) {
-          return bluetoothDevice.id == _bluetoothDeviceIOS?.id;
-        });
-        if (deviceConnectedIndex < 0) {
-          await _bluetoothDeviceIOS?.connect();
-        }
-      }
-
+      final FluetoothDevice fDevice =
+          await Fluetooth().connect(device.id).timeout(timeout);
+      _selectedDevice = fDevice;
       _isConnected = true;
-      selectedDevice?.connected = true;
       return Future<ConnectionStatus>.value(ConnectionStatus.connected);
     } on Exception catch (error) {
-      print('$runtimeType - Error $error');
+      log('$runtimeType - Error $error');
       _isConnected = false;
-      selectedDevice?.connected = false;
+      _selectedDevice = null;
       return Future<ConnectionStatus>.value(ConnectionStatus.timeout);
     }
   }
 
   /// To stop communication between bluetooth device and application
-  Future<ConnectionStatus> disconnect({
-    Duration timeout = const Duration(seconds: 5),
-  }) async {
-    if (Platform.isAndroid) {
-      if (await _bluetoothAndroid?.isConnected ?? false) {
-        await _bluetoothAndroid?.disconnect();
-      }
-      _isConnected = false;
-    } else if (Platform.isIOS) {
-      await _bluetoothDeviceIOS?.disconnect();
-      _isConnected = false;
-    }
-
+  Future<ConnectionStatus> disconnect() async {
+    await Fluetooth().disconnect();
+    _isConnected = false;
+    _selectedDevice = null;
     return ConnectionStatus.disconnect;
   }
 
@@ -123,6 +111,9 @@ class BluePrintPos {
   /// [textScaleFactor] the text scale factor (must be > 0 or null).
   /// note that this currently only works on Android.
   /// defaults to system's font settings.
+  ///
+  /// [batchPrintOptions] to print each [ReceiptSectionText]'s content in batch.
+  /// defaults to [BatchPrintOptions.full].
   Future<void> printReceiptText(
     ReceiptSectionText receiptSectionText, {
     int feedCount = 0,
@@ -135,54 +126,28 @@ class BluePrintPos {
   }) async {
     final int contentLength = receiptSectionText.contentLength;
 
-    final BatchPrintOptions batchOptions;
-    if (batchPrintOptions != null) {
-      batchOptions = batchPrintOptions;
-    } else {
-      if (Platform.isIOS) {
-        batchOptions = const BatchPrintOptions.perNContent(20);
-      } else {
-        batchOptions = BatchPrintOptions.full;
-      }
-    }
+    final BatchPrintOptions batchOptions =
+        batchPrintOptions ?? BatchPrintOptions.full;
 
-    final Iterable<List<int>> startEndIter =
+    final Iterable<List<Object>> startEndIter =
         batchOptions.getStartEnd(contentLength);
 
-    for (final List<int> startEnd in startEndIter) {
+    for (final List<Object> startEnd in startEndIter) {
       final ReceiptSectionText section = receiptSectionText.getSection(
-        startEnd[0],
-        startEnd[1],
+        startEnd[0] as int,
+        startEnd[1] as int,
       );
-      // This is to prevent cropped section at the bottom
-      if (Platform.isIOS) {
-        section.addSpacerPx(15);
-      }
+      final bool isEndOfBatch = startEnd[2] as bool;
       final Uint8List bytes = await contentToImage(
         content: section.getContent(),
         duration: duration,
         textScaleFactor: textScaleFactor,
       );
-      List<int>? croppedExcessTop;
-      // This is to remove the excess top
-      if (Platform.isIOS) {
-        final img.Image? image = img.decodeImage(bytes);
-        if (image != null) {
-          final img.Image cropped = img.copyCrop(
-            image,
-            0,
-            45,
-            image.width,
-            image.height - 45,
-          );
-          croppedExcessTop = img.encodeJpg(cropped);
-        }
-      }
       final List<int> byteBuffer = await _getBytes(
-        croppedExcessTop ?? bytes,
+        bytes,
         paperSize: paperSize,
-        feedCount: feedCount,
-        useCut: useCut,
+        feedCount: isEndOfBatch ? feedCount : batchOptions.feedCount,
+        useCut: isEndOfBatch ? useCut : batchOptions.useCut,
         useRaster: useRaster,
       );
       await _printProcess(byteBuffer);
@@ -218,7 +183,7 @@ class BluePrintPos {
       useRaster: useRaster,
       paperSize: paperSize,
     );
-    _printProcess(byteBuffer);
+    await _printProcess(byteBuffer);
   }
 
   /// This method only for print QR, only pass value on parameter [data]
@@ -230,14 +195,14 @@ class BluePrintPos {
     int size = 120,
     int feedCount = 0,
     bool useCut = false,
-    PaperSize paperSize = PaperSize.mm58,
   }) async {
     final List<int> byteBuffer = await _getQRImage(data, size.toDouble());
-    printReceiptImage(byteBuffer,
-        width: size,
-        feedCount: feedCount,
-        useCut: useCut,
-        paperSize: paperSize);
+    await printReceiptImage(
+      byteBuffer,
+      width: size,
+      feedCount: feedCount,
+      useCut: useCut,
+    );
   }
 
   /// Reusable method for print text, image or QR based value [byteBuffer]
@@ -245,32 +210,14 @@ class BluePrintPos {
   /// But in iOS more complex handler using service and characteristic
   Future<void> _printProcess(List<int> byteBuffer) async {
     try {
-      if (selectedDevice == null) {
-        print('$runtimeType - Device not selected');
-        return Future<void>.value(null);
+      if (!await Fluetooth().isConnected) {
+        _isConnected = false;
+        _selectedDevice = null;
+        return;
       }
-      if (!_isConnected && selectedDevice != null) {
-        await connect(selectedDevice!);
-      }
-      if (Platform.isAndroid) {
-        await _bluetoothAndroid?.writeBytes(Uint8List.fromList(byteBuffer));
-      } else if (Platform.isIOS) {
-        final List<flutter_blue.BluetoothService> bluetoothServices =
-            await _bluetoothDeviceIOS?.discoverServices() ??
-                <flutter_blue.BluetoothService>[];
-        final flutter_blue.BluetoothService bluetoothService =
-            bluetoothServices.firstWhere(
-          (flutter_blue.BluetoothService service) => service.isPrimary,
-        );
-        final flutter_blue.BluetoothCharacteristic characteristic =
-            bluetoothService.characteristics.firstWhere(
-          (flutter_blue.BluetoothCharacteristic bluetoothCharacteristic) =>
-              bluetoothCharacteristic.properties.write,
-        );
-        await characteristic.write(byteBuffer, withoutResponse: true);
-      }
+      await Fluetooth().sendBytes(byteBuffer);
     } on Exception catch (error) {
-      print('$runtimeType - Error $error');
+      log('$runtimeType - Error $error');
     }
   }
 
@@ -294,6 +241,10 @@ class BluePrintPos {
       img.decodeImage(data)!,
       width: customWidth > 0 ? customWidth : paperSize.width,
     );
+    final bool canFullCut = printerHasFeatureOf(
+      _selectedDevice!.name,
+      PrinterFeature.paperFullCut,
+    );
     if (useRaster) {
       bytes += generator.imageRaster(_resize);
     } else {
@@ -302,7 +253,7 @@ class BluePrintPos {
     if (feedCount > 0) {
       bytes += generator.feed(feedCount);
     }
-    if (useCut) {
+    if (useCut && canFullCut) {
       bytes += generator.cut();
     }
     return bytes;