+++ lib/models/models.dart
@@ -1,2 +1,4 @@
-export 'blue_device.dart';
+export 'batch_print_options.dart';
 export 'connection_status.dart';
+export 'printer_feature_rule.dart';
+export 'printer_features.dart';